+++ apps/fabric-website/src/pages/Components/SliderComponentPage.tsx
@@ -7,30 +7,31 @@ const pageStyles: any = require('../PageStyles.module.scss');
 export class SliderComponentPage extends React.Component<any, any> {
   public render(): JSX.Element {
     return (
-      <div className={ pageStyles.basePage }>
+      <div className={pageStyles.basePage}>
         <ComponentPage>
-          <PageHeader pageTitle='Slider' backgroundColor='#038387'
-            links={
-              [
-                {
-                  'text': 'Overview',
-                  'location': 'Overview'
-                },
-                {
-                  'text': 'Best Practices',
-                  'location': 'BestPractices'
-                },
-                {
-                  'text': 'Variants',
-                  'location': 'Variants'
-                },
-                {
-                  'text': 'Implementation',
-                  'location': 'Implementation'
-                }
-              ]
-            } />
-          <SliderPage isHeaderVisible={ false } />
+          <PageHeader
+            pageTitle="Slider"
+            backgroundColor="#038387"
+            links={[
+              {
+                text: 'Overview',
+                location: 'Overview'
+              },
+              {
+                text: 'Best Practices',
+                location: 'BestPractices'
+              },
+              {
+                text: 'Variants',
+                location: 'Variants'
+              },
+              {
+                text: 'Implementation',
+                location: 'Implementation'
+              }
+            ]}
+          />
+          <SliderPage isHeaderVisible={false} />
         </ComponentPage>
       </div>
     );