+++ packages/experiments/src/components/Tile/ShimmerTile/ShimmerTile.types.ts
@@ -1,11 +1,9 @@
 import * as React from 'react';
 import { IStyle } from '../../../Styling';
 import { ISize, IStyleFunctionOrObject } from '../../../Utilities';
-import { TileSize } from '../../Tile';
+import { TileSize } from '../Tile.types';
 
-export interface IShimmerTile {
-
-}
+export interface IShimmerTile {}
 
 /**
  * ShimmerTile component props.