+++ packages/office-ui-fabric-react/src/components/Coachmark/PositioningContainer/PositioningContainer.types.ts
@@ -1,19 +1,15 @@
 import * as React from 'react';
 import { PositioningContainer } from './PositioningContainer';
 import { DirectionalHint } from 'office-ui-fabric-react/lib/common/DirectionalHint';
-import {
-  IPoint,
-  IRectangle
-} from '../../../Utilities';
+import { IPoint, IRectangle } from '../../../Utilities';
 import { IPositionedData } from 'office-ui-fabric-react/lib/utilities/positioning';
 
-export interface IPositioningContainer {
-}
+export interface IPositioningContainer {}
 
 export interface IPositioningContainerTypes extends React.Props<PositioningContainer> {
   /**
-  * All props for your component are to be defined here.
-  */
+   * All props for your component are to be defined here.
+   */
   componentRef?: (component: IPositioningContainer | null) => void;
   /**
    * The target that the positioningContainer should try to position itself based on.