+++ packages/office-ui-fabric-react/src/components/Image/Image.styles.ts
@@ -1,9 +1,5 @@
 import { IImageStyleProps, IImageStyles } from './Image.types';
-import {
-  AnimationClassNames,
-  IStyle,
-  getGlobalClassNames,
-} from '../../Styling';
+import { AnimationClassNames, IStyle, getGlobalClassNames } from '../../Styling';
 
 const GlobalClassNames = {
   root: 'ms-Image',
@@ -14,12 +10,10 @@ const GlobalClassNames = {
   imageCover: 'ms-Image-image--cover',
   imageNone: 'ms-Image-image--none',
   imageLandscape: 'ms-Image-image--landscape',
-  imagePortrait: 'ms-Image-image--portrait',
+  imagePortrait: 'ms-Image-image--portrait'
 };
 
-export const getStyles = (
-  props: IImageStyleProps
-): IImageStyles => {
+export const getStyles = (props: IImageStyleProps): IImageStyles => {
   const {
     className,
     width,
@@ -47,7 +41,7 @@ export const getStyles = (
     transform: 'translate(-50%,-50%)' // @todo test RTL renders transform: translate(50%,-50%);
   };
 
-  return ({
+  return {
     root: [
       classNames.root,
       {
@@ -61,7 +55,7 @@ export const getStyles = (
         }
       ],
       (isCenter || isContain || isCover) && {
-        position: 'relative',
+        position: 'relative'
       },
       className
     ],
@@ -77,10 +71,7 @@ export const getStyles = (
           opacity: 1
         }
       ],
-      isCenter && [
-        classNames.imageCenter,
-        ImageFitStyles
-      ],
+      isCenter && [classNames.imageCenter, ImageFitStyles],
       isContain && [
         classNames.imageContain,
         isLandscape && {
@@ -113,18 +104,21 @@ export const getStyles = (
         }
       ],
       isNotImageFit && [
-        !!width && !height && {
-          height: 'auto',
-          width: '100%'
-        },
-        !width && !!height && {
-          height: '100%',
-          width: 'auto'
-        },
-        !!width && !!height && {
-          height: '100%',
-          width: '100%'
-        }
+        !!width &&
+          !height && {
+            height: 'auto',
+            width: '100%'
+          },
+        !width &&
+          !!height && {
+            height: '100%',
+            width: 'auto'
+          },
+        !!width &&
+          !!height && {
+            height: '100%',
+            width: '100%'
+          }
       ],
       isLoaded && shouldFadeIn && !shouldStartVisible && AnimationClassNames.fadeIn400,
       isLandscape && classNames.imageLandscape,
@@ -133,5 +127,5 @@ export const getStyles = (
       shouldFadeIn && 'is-fadeIn',
       isError && 'is-error'
     ]
-  });
+  };
 };