+++ packages/office-ui-fabric-react/src/components/KeytipLayer/KeytipLayer.styles.ts
@@ -4,18 +4,22 @@ import { hiddenContentStyle, ZIndexes } from '../../Styling';
 
 export const getLayerStyles = (props: ILayerStyleProps): ILayerStyles => {
   return {
-    root: [{
-      // Prioritize the Keytips above all other Layers
-      zIndex: ZIndexes.KeytipLayer
-    }]
+    root: [
+      {
+        // Prioritize the Keytips above all other Layers
+        zIndex: ZIndexes.KeytipLayer
+      }
+    ]
   };
 };
 
 export const getStyles = (props: IKeytipLayerStyleProps): IKeytipLayerStyles => {
   return {
-    innerContent: [{
-      ...hiddenContentStyle,
-      visibility: 'hidden'
-    }]
+    innerContent: [
+      {
+        ...hiddenContentStyle,
+        visibility: 'hidden'
+      }
+    ]
   };
 };