+++ packages/office-ui-fabric-react/src/components/Persona/Persona.tsx
@@ -1,9 +1,5 @@
 import { styled } from '../../Utilities';
-import {
-  IPersonaProps,
-  IPersonaStyleProps,
-  IPersonaStyles
-} from './Persona.types';
+import { IPersonaProps, IPersonaStyleProps, IPersonaStyles } from './Persona.types';
 import { PersonaBase } from './Persona.base';
 import { getStyles } from './Persona.styles';
 
@@ -11,7 +7,4 @@ import { getStyles } from './Persona.styles';
  * Personas are used for rendering an individual's avatar, presence and details.
  * They are used within the PeoplePicker components.
  */
-export const Persona = styled<IPersonaProps, IPersonaStyleProps, IPersonaStyles>(
-  PersonaBase,
-  getStyles
-);
+export const Persona = styled<IPersonaProps, IPersonaStyleProps, IPersonaStyles>(PersonaBase, getStyles);