+++ packages/experiments/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log - @uifabric/experiments
 
-This log was last generated on Fri, 01 Jun 2018 10:18:43 GMT and should not be manually modified.
+This log was last generated on Tue, 05 Jun 2018 00:44:30 GMT and should not be manually modified.
+
+## 6.1.2
+Tue, 05 Jun 2018 00:44:30 GMT
+
+### Patches
+
+- Added Prettier
+
+## 6.1.1
+Mon, 04 Jun 2018 10:16:13 GMT
+
+### Patches
+
+- Updating react typings.
+- Fixed tests to be compatible with jest 23
 
 ## 6.1.0
 Fri, 01 Jun 2018 10:18:43 GMT